chore: add small cards carousel
