+++ components/FindResidents/Resident/ResidentPage/Forms/index.js
@@ -8,29 +8,29 @@ import {
 
 import { theme } from '../../../../../modules/theme';
 import ComingSoonSVG from '../../../../../assets/graphics/static/Adventurer.svg';
-import SmallCards from '../../../../Cards/SmallCards';
+import SmallCardsCarousel from '../../../../Cards/SmallCardsCarousel';
 
 const Forms = ({
   puenteForms, navigateToNewRecord, surveyee, setView
 }) => (
-  <View style={styles.container}>
-    <Title style={styles.title}>Completed Forms</Title>
-    <Text style={styles.category}>Form Category</Text>
-    <View
-      style={styles.horizontalLine}
-    />
-    <ComingSoonSVG width={200} height={200} />
-    <Text>Coming Soon</Text>
-    <Title style={styles.title}>Suggested Forms</Title>
-    <SmallCards
-      puenteForms={puenteForms}
-      navigateToNewRecord={navigateToNewRecord}
-      surveyee={surveyee}
-      setView={setView}
-      setUser
-    />
-  </View>
-);
+    <View style={styles.container}>
+      <Title style={styles.title}>Completed Forms</Title>
+      <Text style={styles.category}>Form Category</Text>
+      <View
+        style={styles.horizontalLine}
+      />
+      <ComingSoonSVG width={200} height={200} />
+      <Text>Coming Soon</Text>
+      <Title style={styles.title}>Suggested Forms</Title>
+      <SmallCardsCarousel
+        puenteForms={puenteForms}
+        navigateToNewRecord={navigateToNewRecord}
+        surveyee={surveyee}
+        setView={setView}
+        setUser
+      />
+    </View>
+  );
 
 const styles = StyleSheet.create({
   container: {
