+++ components/FindResidents/index.js
@@ -76,7 +76,7 @@ const FindResidents = ({
         </>
       )}
 
-      {query !== '' && filterList(residents).map((listItem,) => (
+      {filterList(residents).map((listItem,) => (
         <View key={listItem.objectId}>
           <ResidentCard
             resident={listItem}
